Deduplicate ps spawn: move collectProcessInfo to port-scanner with TTL cache
Both pollers (port-scanner and resource-monitor) previously called ps independently
on every 10s cycle. Now collectProcessInfo lives in port-scanner and is shared via
a 5s TTL cache — whichever poller fires first pays for the spawn, the other gets
the cached result.

Also replaces the inline BFS in resource-monitor with collectDescendants() from
port-scanner, removing the duplicated tree-walk logic.
